+++ RELEASE_NOTES.md
@@ -2,7 +2,7 @@
 
 ## New in git-machete 3.17.0
 
-- added: `<branch>` parameter to `git machete github retarget-pr` command
+- added: `--branch=<branch>` option to `git machete github retarget-pr` command
 
 ## New in git-machete 3.16.3